Refactor remeshing into free functions
