+++ src/pmp/algorithms/remeshing.cpp
@@ -1,7 +1,7 @@
 // Copyright 2011-2020 the Polygon Mesh Processing Library developers.
 // Distributed under a MIT-style license, see LICENSE.txt for details.
 
-#include "pmp/algorithms/Remeshing.h"
+#include "pmp/algorithms/remeshing.h"
 
 #include <cmath>
 
@@ -16,6 +16,71 @@
 #include "pmp/Timer.h"
 
 namespace pmp {
+
+class Remeshing
+{
+public:
+    Remeshing(SurfaceMesh& mesh);
+
+    void uniform_remeshing(Scalar edge_length, unsigned int iterations = 10,
+                           bool use_projection = true);
+
+    void adaptive_remeshing(Scalar min_edge_length, Scalar max_edge_length,
+                            Scalar approx_error, unsigned int iterations = 10,
+                            bool use_projection = true);
+
+private:
+    void preprocessing();
+    void postprocessing();
+
+    void split_long_edges();
+    void collapse_short_edges();
+    void flip_edges();
+    void tangential_smoothing(unsigned int iterations);
+    void remove_caps();
+
+    Point minimize_squared_areas(Vertex v);
+    Point weighted_centroid(Vertex v);
+
+    void project_to_reference(Vertex v);
+
+    bool is_too_long(Vertex v0, Vertex v1) const
+    {
+        return distance(points_[v0], points_[v1]) >
+               4.0 / 3.0 * std::min(vsizing_[v0], vsizing_[v1]);
+    }
+    bool is_too_short(Vertex v0, Vertex v1) const
+    {
+        return distance(points_[v0], points_[v1]) <
+               4.0 / 5.0 * std::min(vsizing_[v0], vsizing_[v1]);
+    }
+
+    SurfaceMesh& mesh_;
+    std::shared_ptr<SurfaceMesh> refmesh_;
+
+    bool use_projection_;
+    std::unique_ptr<TriangleKdTree> kd_tree_;
+
+    bool uniform_;
+    Scalar target_edge_length_;
+    Scalar min_edge_length_;
+    Scalar max_edge_length_;
+    Scalar approx_error_;
+
+    bool has_feature_vertices_{false};
+    bool has_feature_edges_{false};
+    VertexProperty<Point> points_;
+    VertexProperty<Point> vnormal_;
+    VertexProperty<bool> vfeature_;
+    EdgeProperty<bool> efeature_;
+    VertexProperty<bool> vlocked_;
+    EdgeProperty<bool> elocked_;
+    VertexProperty<Scalar> vsizing_;
+
+    VertexProperty<Point> refpoints_;
+    VertexProperty<Point> refnormals_;
+    VertexProperty<Scalar> refsizing_;
+};
 
 Remeshing::Remeshing(SurfaceMesh& mesh)
     : mesh_(mesh), refmesh_(nullptr), kd_tree_(nullptr)
@@ -860,6 +925,21 @@ Point Remeshing::weighted_centroid(Vertex v)
     p /= ww;
 
     return p;
+}
+
+void uniform_remeshing(SurfaceMesh& mesh, Scalar edge_length,
+                       unsigned int iterations, bool use_projection)
+{
+    Remeshing(mesh).uniform_remeshing(edge_length, iterations, use_projection);
+}
+
+void adaptive_remeshing(SurfaceMesh& mesh, Scalar min_edge_length,
+                        Scalar max_edge_length, Scalar approx_error,
+                        unsigned int iterations, bool use_projection)
+{
+    Remeshing(mesh).adaptive_remeshing(min_edge_length, max_edge_length,
+                                       approx_error, iterations,
+                                       use_projection);
 }
 
 } // namespace pmp
