+++ tests/RemeshingTest.cpp
@@ -3,7 +3,7 @@
 
 #include "gtest/gtest.h"
 
-#include "pmp/algorithms/Remeshing.h"
+#include "pmp/algorithms/remeshing.h"
 #include "pmp/algorithms/Features.h"
 #include "pmp/algorithms/shapes.h"
 #include "pmp/algorithms/triangulation.h"
@@ -20,9 +20,9 @@ TEST(RemeshingTest, adaptive_remeshing_with_features)
     triangulate(mesh);
     Features(mesh).detect_angle(25);
     auto bb = bounds(mesh).size();
-    Remeshing(mesh).adaptive_remeshing(0.001 * bb,  // min length
-                                       1.0 * bb,    // max length
-                                       0.001 * bb); // approx. error
+    adaptive_remeshing(mesh, 0.001 * bb, // min length
+                       1.0 * bb,         // max length
+                       0.001 * bb);      // approx. error
     EXPECT_EQ(mesh.n_vertices(), 6u);
 }
 
@@ -31,9 +31,9 @@ TEST(RemeshingTest, adaptive_remeshing_with_boundary)
     // mesh with boundary
     auto mesh = open_cone();
     auto bb = bounds(mesh).size();
-    Remeshing(mesh).adaptive_remeshing(0.01 * bb,  // min length
-                                       1.0 * bb,   // max length
-                                       0.01 * bb); // approx. error
+    adaptive_remeshing(mesh, 0.01 * bb, // min length
+                       1.0 * bb,        // max length
+                       0.01 * bb);      // approx. error
     EXPECT_EQ(mesh.n_vertices(), size_t(65));
 }
 
@@ -48,15 +48,15 @@ TEST(RemeshingTest, adaptive_remeshing_with_selection)
             selected[v] = true;
 
     auto bb = bounds(mesh).size();
-    Remeshing(mesh).adaptive_remeshing(0.01 * bb,  // min length
-                                       1.0 * bb,   // max length
-                                       0.01 * bb); // approx. error
+    adaptive_remeshing(mesh, 0.01 * bb, // min length
+                       1.0 * bb,        // max length
+                       0.01 * bb);      // approx. error
     EXPECT_EQ(mesh.n_vertices(), size_t(62));
 }
 
 TEST(RemeshingTest, uniform_remeshing)
 {
     auto mesh = open_cone();
-    Remeshing(mesh).uniform_remeshing(0.5);
+    uniform_remeshing(mesh, 0.5);
     EXPECT_EQ(mesh.n_vertices(), size_t(41));
 }
